# src/Server/wwwroot/swagger/v1/swagger.json
{
  "x-generator": "NSwag v13.18.2.0 (NJsonSchema v10.8.0.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "timeNet aAPI",
    "version": "1.0.0"
  },
  "paths": {
    "/api/config": {
      "get": {
        "tags": [
          "Config"
        ],
        "operationId": "Config_GetConfig",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Config"
        ],
        "operationId": "Config_SetConfig",
        "requestBody": {
          "x-name": "config",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigDTO"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/css": {
      "get": {
        "tags": [
          "Config"
        ],
        "operationId": "Config_GetConfigCss",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/config/css/file": {
      "get": {
        "tags": [
          "Config"
        ],
        "operationId": "Config_GetConfigCssFile",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/developer/seed": {
      "put": {
        "tags": [
          "Developer"
        ],
        "operationId": "Developer_Seed",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/developer/reset": {
      "put": {
        "tags": [
          "Developer"
        ],
        "operationId": "Developer_Reset",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/developer/notification": {
      "post": {
        "tags": [
          "Developer"
        ],
        "operationId": "Developer_SendNotification",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "error",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 2
          },
          {
            "name": "seconds",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 3
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/intervals/start": {
      "get": {
        "tags": [
          "Intervals"
        ],
        "operationId": "Intervals_GetIntervalsByStartTimes",
        "parameters": [
          {
            "name": "starts",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string",
                "format": "date-time"
              }
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntervalDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/intervals/id": {
      "get": {
        "tags": [
          "Intervals"
        ],
        "operationId": "Intervals_GetIntervalsByIds",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "integer",
                "format": "int32"
              }
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntervalDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/intervals": {
      "get": {
        "tags": [
          "Intervals"
        ],
        "operationId": "Intervals_GetIntervalsByQuery",
        "parameters": [
          {
            "name": "IncludeActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "x-position": 1
          },
          {
            "name": "Earliest",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "Latest",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "Tags",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string"
              }
            },
            "x-position": 4
          },
          {
            "name": "IsAnnotated",
            "in": "query",
            "schema": {
              "type": "boolean",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "Exclude",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "integer",
                "format": "int32"
              }
            },
            "x-position": 6
          },
          {
            "name": "Updated",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 7
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 8
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 9
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponseOfIntervalDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Intervals"
        ],
        "operationId": "Intervals_ApplyInterval",
        "requestBody": {
          "x-name": "input",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntervalDTO"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Intervals"
        ],
        "operationId": "Intervals_DeleteIntervals",
        "requestBody": {
          "x-name": "ids",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/intervals/latest": {
      "get": {
        "tags": [
          "Intervals"
        ],
        "operationId": "Intervals_GetLatest",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntervalDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/intervals/next": {
      "get": {
        "tags": [
          "Intervals"
        ],
        "operationId": "Intervals_GetNext",
        "parameters": [
          {
            "name": "time",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntervalDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/intervals/prev": {
      "get": {
        "tags": [
          "Intervals"
        ],
        "operationId": "Intervals_GetPrevious",
        "parameters": [
          {
            "name": "time",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntervalDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/intervals/import": {
      "post": {
        "tags": [
          "Intervals"
        ],
        "operationId": "Intervals_ImportIntervals",
        "requestBody": {
          "x-name": "data",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/IntervalDTO"
                }
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/statistics/heatmap": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "operationId": "Statistics_GetHeatmap",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "number",
                    "format": "double"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/statistics/toptags": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "operationId": "Statistics_GetTopTags",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/statistics/misc": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "operationId": "Statistics_GetMisc",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiscStats"
                }
              }
            }
          }
        }
      }
    },
    "/api/tags/name": {
      "get": {
        "tags": [
          "Tags"
        ],
        "operationId": "Tags_GetTagsByNames",
        "parameters": [
          {
            "name": "names",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string"
              }
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tags/id": {
      "get": {
        "tags": [
          "Tags"
        ],
        "operationId": "Tags_GetTagsByIds",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "integer",
                "format": "int32"
              }
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tags": {
      "get": {
        "tags": [
          "Tags"
        ],
        "operationId": "Tags_GetTagsByQuery",
        "parameters": [
          {
            "name": "Name",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "Exclude",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string"
              }
            },
            "x-position": 2
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponseOfTagDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tags"
        ],
        "operationId": "Tags_ApplyTag",
        "requestBody": {
          "x-name": "tag",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagDTO"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tags"
        ],
        "operationId": "Tags_DeleteTags",
        "requestBody": {
          "x-name": "ids",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConfigDTO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TrackedEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "theme": {
                "type": "string"
              },
              "highlight": {
                "type": "string"
              },
              "colours": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ColourMatch"
                }
              },
              "updated": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "ColourMatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "regex": {
            "type": "string"
          },
          "colour": {
            "type": "string"
          }
        }
      },
      "TrackedEntity": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "updated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "IntervalDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "annotation": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "QueryResponseOfIntervalDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntervalDTO"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TagDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "intervals": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MiscStats": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "totalTags": {
            "type": "integer",
            "format": "int32"
          },
          "totalIntervals": {
            "type": "integer",
            "format": "int32"
          },
          "oldest": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/IntervalDTO"
              }
            ]
          },
          "longest": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/IntervalDTO"
              }
            ]
          }
        }
      },
      "QueryResponseOfTagDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDTO"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}